creat api add user

+++ src/main/java/com/example/itspower/repository/repositoryjpa/UserJPARepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserJPARepository extends JpaRepository<UserEntity,Long> {
+public interface UserJPARepository extends JpaRepository<UserEntity, Integer> {
     UserEntity findByUserName(String userName);
 }
